Convert DistributionMirror queries to Storm style
This prepares the way for converting the corresponding models from
SQLObject to Storm syntax.

Converting manual SQL to compiled Storm queries wasn't necessary in
every case here, but we generally prefer it where possible, and in some
cases the act of doing this exposed some unnecessarily complicated
queries: the `DistributionMirror` joins in
`DistributionMirror.getSummarizedMirroredSourceSeries` and
`DistributionMirror.getSummarizedMirroredArchSeries` weren't needed, and
`DistributionMirrorSet.getMirrorsToProbe` can avoid a query in some
cases.