Disable verifyFormat check for .debs and .ddebs
The `process-upload` script that runs on build uploads is close to
scaling limits: at the moment it doesn't always manage to outpace
`buildd-manager` adding items to its queue.  It's therefore worth some
optimization effort.

For large .deb uploads, especially .ddebs, the `verifyFormat` and
`verifyDebTimestamp` can both be very slow (minutes) and have some
overlap: the former runs `dpkg-deb` and the latter uses
`apt_inst.DebFile`, but both iterate over the control and data parts of
the package.  `verifyDebTimestamp` is effectively a superset of
`verifyFormat` for all practical purposes, so let's avoid running both.