Tell DiskPool about source version when adding/removing files
Some repository formats, such as Python indexes, organize files by
version as well as by name.  To accommodate this, tell the disk pool
about the source package version as well as the source package name when
adding or removing files.  The Artifactory pool implementation also
needs to store this in a property so that it can construct an entry for
the file when updating its properties.