Set explicit webservice versions for lp.bugs
`lp._schema_circular_imports` has historically grandparented in explicit
webservice version declarations for exports that predated the
requirement for explicit version declarations.  It's long past time to
deal with these properly, so start by moving the declarations for
entries and collections under `lp.bugs` to the proper places.

Comparing the WADL files before and after this change is difficult since
they don't have a stable sort order, but they stay the same size which
is a good indication that the set of exported items hasn't changed.