Emit statsd metric when a cron script is disabled
We sometimes disable cron scripts for certain kinds of deployments, in
particular schema updates.  This can cause spurious alerts, because the
usual timing metrics we emit when scripts complete aren't emitted for
the duration of the deployment; this deters us from having
appropriately-sensitive alerts, and means we have to micromanage the
timing of certain deployments unnecessarily.

To avoid this problem, emit a zero-duration timing metric when a cron
script exits early due to being disabled.  The trade-off here is that
scripts will take somewhat longer to exit in this situation since they
need to load the Zope component architecture first, but that's not a
serious problem.