Clarify confusing comments around restricted processors
Comments in several tests used the terms "enabled" and "disabled" to
talk almost interchangeably about both (1) whether an archive or recipe
is allowed to dispatch builds for a given processor and (2) whether the
corresponding UI checkbox is greyed out or not.  This could very easily
cause confusion, so rewrite these comments to be more explicit about
what they mean.