Fix publishing of isolated binaries
When we upload a CI build to an archive, the resulting
`BinaryPackagePublishingHistory` rows have no associated
`SourcePackageRelease`, which causes publishing to fail since it uses
the name and version from the SPR to compute paths in the pool.

Use the name and version from the `BinaryPackageRelease` instead in this
case.  This is quite awkward, and we may need to do something different
with CI builds in the longer term (perhaps storing a name and version on
`CIBuild` instead?), but it does the job for now.