Extend filetype detection in CIBuildUploadJob
`BinaryPackageRelease.addFile` attempts to detect the
`BinaryPackageFileType` based on the file name, but this isn't likely to
be reliably possible for all new package types.  Determine this in
`CIBuildUploadJob` and pass it down.

The intent of the distinction between `BinaryPackageFormat` and
`BinaryPackageFileType` is unclear to me, but their values don't
completely coincide so merging them would be non-trivial.  For now, we
can at least maintain a one-to-one mapping for the values relevant to
`CIBuildUploadJob`.