Eliminate hardcoded line number in test_stacktrace
The hardcoded line number in `lp.services.tests.test_stacktrace` that
must match the source line number on which it appears has bitten me
twice in the past while refactoring, and it's an obstacle to being able
to run source code reformatting tools reliably.  Use `ast` to detect
that line number dynamically instead.