Handle publishing to the Conda archive layout
This required some refactoring, as the directory where a Conda package
file should be published was difficult to determine using only
information that the pool implementation already had: it would have had
to derive the Conda architecture name from the Debian architecture name,
which would have been fragile and difficult to maintain.

Instead, parse the `subdir` entry from the Conda index, and pass that
down via `BinaryPackageRelease.user_defined_fields`.  This requires
ensuring that the pool has access to the `BinaryPackageRelease` (via
`IPackageReleaseFile`).  In one case that's difficult:
`DiskPool.pathFor` is called by `FTPArchiveHandler.publishFileLists`,
and getting hold of the whole release file row in that case would
require significantly widening a query with an already large result set.
Fortunately, it's straightforward in every other non-test case, so I was
able to arrange for `DiskPool.pathFor` to accept either an
`IPackageReleaseFile` or a file name with only a small amount of
ugliness, and everything else derives the file name from the
`IPackageReleaseFile` where needed.

`Publisher.C_updateArtifactoryProperties` also needed a little work.  It
now needs to pass the appropriate `IPackageReleaseFile`, but the current
code could only find that if that file was still published anywhere, and
not if all those publications had been deleted but the file was still in
the pool pending removal.  Use an alternative approach to get at the
release files in such cases.