Allow CI builds to be uploaded multiple times to different targets
It's quite confusing that only one `CIBuildUploadJob` can ever be run
for a given CI build, even if you want to release it to multiple
channels or multiple archives or similar; this is because there's a
unique constraint on `BinaryPackageRelease (ci_build,
binarypackagename)`.  Lift this restriction by just creating new
`BinaryPackagePublishingHistory` rows for the relevant
`BinaryPackageRelease` in that case.