Upgrade Markdown and sanitize its output using bleach
This makes a start at reviving the old experiment to allow user-supplied
Markdown.  `safe_mode` is deprecated, so use `bleach` to sanitize the
resulting HTML instead (important to avoid trivial XSS attacks), and
also add the `fenced_code` extension since that looks like something
we're likely to want.

As per the linked bug, there are a number of things still to fix, and
this is still behind a feature flag.

LP: #391780