Add BPR (ci_build, binarypackagename, binpackageformat) index
We previously had a unique index on `BPR (ci_build, binarypackagename)`.
However, this doesn't work well with multiple binary package formats in
play: the same CI build might well build the same package as both a
wheel and a Conda package.  This hot patch adds a narrower index which
also includes the format; once it's been applied, we can drop the
broader index.