Fix quoting of semicolons in Artifactory properties
Despite what the Artifactory documentation says, semicolons in
properties must be quoted in order to avoid being confused with property
separators.  The exact observed behaviour of the implementation doesn't
seem to be documented anywhere I've found, and it's quite strange in
places, but I've done the best I could to make the test fixture agree
with reality as closely as possible.