URL-quote property values for ArtifactoryPath.deploy_file
`ArtifactoryPath.deploy_file` and `ArtifactoryPath.set_properties` have
two entirely distinct quoting conventions.  Astonishingly, the Python
bindings for the former leave this entirely up to the caller without
saying so.  Fortunately, all that `deploy_file` needs is URL-quoting
according to `urllib.parse.quote_plus`.  Do this.

In the process of fixing this, I noticed that the Artifactory fixture's
endpoints don't correctly unquote the URL path, so I fixed that.