Fix transaction management error when failing a job
Many of Launchpad's jobs are structured using a separate table (e.g.
`ArchiveJob`) that has a reference to a `Job` row, so evaluating
`self.job` may involve an SQL query if that reference is not in the
Storm cache (as is the case after any commit or abort, since both of
those invalidate the cache).  This means that a job that executes an SQL
statement that raises an error may crash as follows when trying to mark
the job as failed:

    storm.database.InFailedSqlTransaction: current transaction is
    aborted, commands ignored until end of transaction block

To fix this, take slightly more control of transaction management.
`self.job.fail(manage_transaction=True)` aborts the transaction at the
start of the method and commits the transaction at the end, but calling
that in the context of an already-aborted transaction requires
`self.job` to be in the Storm cache.  Instead, lift the abort/commit out
by one layer so that `self.job` is evaluated in a new transaction.