Avoid using simplejson with lazr.restful.ResourceJSONEncoder
Now that `ResourceJSONEncoder` is based on `json.JSONEncoder` rather
than on `simplejson.encoder.JSONEncoderForHTML`, it no longer supports
extra parameters that `simplejson.dumps` wants to pass to it, so use it
via the ordinary `json.dumps`.

(In hindsight, lazr.restful 2.0.1 should probably have been 2.1.0 or
even 3.0.0 due to this.  But it's unlikely to be worth going back and
fixing that now - Launchpad is almost certainly the only significant
user of this code.)