Remove development builder proxy configuration
The development configuration for the builder proxy is somewhat
convenient if you're actually working on the builder proxy.  In other
situations, it's an obstacle: it's one more obscure piece of
configuration that you typically need to remove if you're trying to set
up a local builder.  Local builders generally don't have the sort of
strict DNS/networking configuration that we run with in production and
that requires this proxy, so the proxy is normally of little benefit
there.

Let's just remove this configuration to simplify the setup procedure,
and enable it only for the tests that need it; people who need to set up
a local builder proxy as well should have little difficulty configuring
Launchpad to use it.