Refactor SourcePackageRelease creation for CI builds
DB patch 2210-44-0 added support for `SourcePackageRelease` rows based
on `CIBuild`s, but that wasn't yet reflected in the ORM model.  Add the
new `ci_build` column, and rearrange `SourcePackageRelease` and
`SourcePackagePublishingHistory` creation methods so that dpkg-specific
arguments are optional.

(This is a clear case where `black`'s formatting would have made the
diff more concise, but we haven't got as far as `lp.registry` and
`lp.soyuz` there yet.)