Only allow uploading CI builds for package namespaces
`Archive.uploadCIBuild` previously allowed uploads of CI builds from Git
repositories in any namespace.  Constrain this to only allowing uploads
from distribution source package namespaces.  This has the benefit that
it means we always know the source package name, which allows checking
finer-grained archive permissions and will shortly make it easier to
upload source package files as well.

I don't believe that this will be a problem for any of the relatively
small number of ways in which people are currently using this API on
production.