Create a source publication when uploading a CI build
This has some advantages and disadvantages, but on the whole I think the
advantages outweigh the disadvantages.

The downsides are:

 * if the CI build produces nothing that looks like source code at all,
   then we end up with a slightly odd "source package" with no files
   attached to it;

 * if builds do produce something that looks like source code (e.g. a
   Python sdist), then they're on the honour system to either produce it
   from only one architecture or to produce something equivalent from
   all architectures.

On the other hand:

 * Launchpad's data model for packages makes the assumption in many
   places that binary packages have an associated source package, and my
   attempts to make it tolerate isolated binary packages have only been
   marginally successful;

 * treating source-like artifacts (Python sdists, zip files of Go source
   code, etc.) as source packages rather than binary packages makes more
   sense to humans;

 * source packages provide a useful grouping for collections of binary
   packages from builds of the same version;

 * creating source publications means that the PPA web UI works for PPAs
   populated from CI builds with only a few small adjustments (not in
   this branch);

 * creating source publications means that it should be straightforward
   to get the existing package copying mechanism to work.

I've slightly cheekily renamed `SourcePackageType.SDIST` to
`SourcePackageType.CI_BUILD`.  The type of an individual source file is
represented separately using `SourcePackageFileType`, and since the data
model's assumptions generally involve a single source with zero or more
binaries, it's simpler for `SourcePackageRelease.format` not to need to
change depending on the types of the individual files it contains.
There was previously no way to create a source package with
`format=SDIST` outside the test suite, and a query on staging returns
zero rows.

As a concrete example of a source package file type and in order to
allow more useful testing, I've added support for uploading Python
sdists from CI builds as part of this branch.