Inline relevant parts of zope.session
`zope.session` has some unfortunate dependencies on things like `ZODB`
which it would be helpful to be able to avoid.  It turns out that we
don't actually use very much of its code, because we substitute a
PostgreSQL-based session implementation for most of it; and for the
parts we do use, we often have to work around its behaviour in
inconvenient ways.

This branch inlines the relevant bits of `zope.session`, cut down
heavily to only what Launchpad needs.  It's still unnecessarily
complicated, but this is about as far as I can reduce it without causing
behaviour changes, which ought to be in a separate branch.

`zope.session` is still present in the virtualenv after this change due
to dependencies from `zope.app.appsetup` and `zope.app.wsgi`, but
removing the direct dependency was the hardest part.