Remove direct uses of zope.container
Launchpad used `zope.container.testing` without a direct dependency on
`zope.container`, which is clearly wrong.  The point of this seems to
have been to run some tests with the Zope component architecture set up;
a clearer and more efficient way to do this is to put the relevant tests
in `FunctionalLayer`.