Have "make schema" apply DB patches in separate sessions
This is temporarily necessary because of the PL/Python 2 to 3 migration:
PostgreSQL refuses to have both old and new functions referenced in the
same session, which happens if we try to apply the full 2210-* patch
sequence at once.  It's a bit awkward because there's no obvious way to
make `--dry-run` work in conjunction with this mode, but we should be
able to revert this after rebaselining.