Integrate material from README better into documentation
The `README` file consisted of a few different conceptual pieces:

 * Introductory material summarizing the project, suitable for the top
   of the developer documentation's home page;
 * Brief remarks on who runs the project;
 * A "Getting started" section;
 * Explanatory material on navigating the code tree.

Move most of these to their proper positions in the developer
documentation.  I think it makes sense to maintain at least a brief
text-only `README` file for the benefit of people who clone the
repository directory or look at it on git.launchpad.net, but I've cut it
down and updated some of its links.