Rename NameBlacklist table to NameBlocklist
I've left behind aliases (in the form of a view that meets the
conditions in
https://www.postgresql.org/docs/10/sql-createview.html#SQL-CREATEVIEW-UPDATABLE-VIEWS
and a couple of simple functions) to keep the old names working until
they're no longer referenced in code.  I've also only changed the table
names in sampledata and not the data itself for now, for the same
reason.

The replaced functions are just the result of replacing "blacklist" with
"blocklist" in the function bodies.