Refactor metadata scanners to allow for multiple files
Some packaging formats consist of multiple files.  For example, Debian
source packages always involve a `.dsc` control file and at least one
other file containing the actual code; and Go modules seem to consist of
`.info` and `.mod` files along with the zip file that contains the
actual code.

To allow for this, the metadata scanning arrangements in
`CIBuildUploadJob` need to have a slightly different structure: each
scanner needs to be able to see all the available files and potentially
consume more than one of them in one go.

I also converted this code to `pathlib` in passing, which is a little
nicer.