Implement metadata scanning for Go modules
This also requires parsing `go.mod` files.  There doesn't seem to be an
existing parser for these in Python, but the format is relatively
well-specified, so I implemented a simple one using `pyparsing`.  It
doesn't pretend to handle the whole file format, just enough for what we
need.

We store the module path in `SourcePackageRelease.user_defined_fields`,
which is as expected by `lp.archivepublisher.artifactory`.