Simplify flake8 configuration
Remove Python 2 builtins and all the whitespace-only ignore directives,
now that we're using `black` so that it's pretty easy to comply with
those.  I fixed up some remaining violations, including a small amount
of manual adjustment in the handful of files that still require Python 2
support so that we can't yet run `black` properly.