Remove FAQ and redistribute its contents
Daniele Procida's opinion was that FAQs are an antipattern in
documentation, essentially (I think) because they tend not to actually
be what's frequently asked.  It would make more sense to distribute the
material here into other relevant pages anyway, so do so.

I omitted "What version of Python is required?" entirely, since that's
already covered near the top of `doc/how-to/running.rst`.