Return proxied objects from various Translations factory methods
`LaunchpadObjectFactory` issues `UnproxiedFactoryMethodWarning` when its
methods return objects not wrapped in a security proxy, since that tends
to result in tests that are less accurate simulations of production.

Some tests failed in a way that revealed that `findIdenticalMessage`
wasn't on the `ITranslationMessage` interface, which I think is a bug
since it was called by production code outside the model
(`lp.translations.utilities.translationmerger`), so I added that and
also implemented it on `PlaceholderTranslationMessage`.  This allowed
removing a `removeSecurityProxy` call from
`lp.translations.utilities.translationmerger`, which seems like a good
thing for hygiene.