Publish Python sdists to the correct paths
When creating a `SourcePackageRelease` from a `CIBuild`, we have to take
the release's source package name from the namespace of the build's Git
repository, since that means everything is grouped together properly
under the same source package name, and it also allows us to handle
source package types whose true package name can't be used as a source
package name (notably Go modules); teams building these also generally
try to use the same source package name as used in Ubuntu, to make it
easier to associate fixes in different distributions.

However, this meant that we published Python sdists to the wrong paths
if their true package name didn't correspond to the repository
namespace.  To fix this, pass through the true package name via
`SourcePackageRelease.user_defined_fields` and use that in the
publishing machinery.

LP: #1983401