Fix handling of MaloneView.error_message
Commit 65e983ad01a81e2ec3d6ceb0bb08bf7868b67e7f changed
`MaloneView.error_message` from being initialized to None to being a
property that returns None, due to needing to convince `mypy` that its
type was compatible with that of `LaunchpadView.error_message`.
However, this caused test failures in
`lp.bugs.browser.tests.test_bugs.TestMaloneView.test_redirect_unknown_bug_fail`
and
`lp.bugs.browser.tests.test_bugs.TestMaloneView.test_redirect_list_of_bug_fail`,
because `MaloneView._redirectToBug` expected to be able to assign to
`error_message`.

I think it's best to rearrange `MaloneView` to conform to the
expectations of the base class for `error_message`: that is, assignments
to it must be either None or an `IStructuredString`.  The `structured`
helper makes it easier to write pages in a way that avoids XSS attacks
by separating fixed text from interpolated variables such that
interpolations are HTML-escaped.  As it happens I satisfied myself that
this page was already XSS-safe, but this approach makes it a bit clearer
in a way that's more usual for Launchpad, and makes it easier to keep
the type-checker happy.