Fix logic in POTemplateEditView._return_url
To satisfy `mypy`, commit 65e983ad01a81e2ec3d6ceb0bb08bf7868b67e7f
changed the definition of a couple of properties in `LaunchpadFormView`
from:

    next_url = _return_url
    cancel_url = _return_url

... to:

    @property
    def next_url(self):
        return self._return_url

    @property
    def cancel_url(self):
        return self._return_url

This looks fairly innocuous, but it has an interesting side-effect: it
means that redefining `_return_url` in subclasses is now effective, when
previously it wasn't because the `next_url` and `cancel_url` names were
bound to the definition in the superclass.  As a result, I don't think
the property in the subclass (`POTemplateEditView`) was being called in
practice.  Arranging for it to be called caused failures in
`lib/lp/translations/stories/standalone/xx-potemplate-admin.rst` and
`lib/lp/translations/stories/standalone/xx-potemplate-edit.rst`.  I
considered simply removing the property in the subclass, but based on
the comment at the top there does seem to be a point to it, so I made
the effort to repair it.

The problem was that `POTemplateEditView._return_url` failed to check
whether it needs to construct a new URL in the case where the default
return URL is provided by a hidden input in the form; but that's
normally the case when an edit form is being submitted.  Moving the
check for whether the default return URL contains "/+pots/" out one
level, so that it's called regardless of the source of the default
return URL, fixes the problem.

I also removed `POTemplateEditView._return_attribute_name`, since it was
only used by `LaunchpadFormView._return_url` which is now overridden
here, so it had become effectively dead code.