Rename `apt_repositories` into `package_repositories`
Also rename the corresponding helper function.

This needs to be a two-phase process. Once `launchpad-buildd` works
with the new naming, the deprecated occurrences with `apt_repositories`
can be removed.

This is not only about renaming. This is also a preparation for
switching from replacing repositories with adding additional
repositories further down the stack in `launchpad-buildd` respectively
`lpraft`.