Avoid writing to code tree for config-specific ZCML overrides
Launchpad configs often include some ZCML overrides (for example,
`configs/testrunner/yuitest.zcml` hooks up a fixture for JavaScript
testing).  These were previously loaded by way of service startup code
that generated `zcml/+config-overrides.zcml`, which was then loaded by
`zcml/webapp.zcml` and `zcml/ftesting.zcml`.

However, this arrangement required that service startup code be able to
write to the code tree.  This didn't work well in the context of Juju
charms based on the `ols` layer (which arrange for the code tree to be
writable only by root), and it essentially amounts to self-modifying
code which seems like dubious design in general.

To avoid this problem, implement a new `lp:includeLaunchpadOverrides`
ZCML directive which includes all the `*.zcml` files in the current
instance's configuration.  This is equivalent to the old arrangement,
but without requiring the service to modify its own code tree at
startup.