UCT import: use `Packaging` model for `upstream`
Previously, we were trying to find an LP `Product` with exactly the same name as a package in UCT record to handle the status of `upstream`.

Now, we use the `Packaging` model to locate a `Product` by package name.

More specifically, we use the `DistributionSourcePackage.upstream_product` property for that.