Adjust lp.services.mime for Python >= 3.8
Python 3.8 slightly changed how the `mimetypes` module is initialized
(https://github.com/python/cpython/pull/14375), and this interacted
badly with the reinitialization code in `zope.contenttype` in such a way
as to make the extensions in `lp.services.mime` be ineffective.  Work
around this in a way that works with both old and new versions of Python
by monkey-patching `mimetypes.init`.