Adjust GitHostingClient blob encoding tests for Python >= 3.8
https://github.com/python/cpython/pull/7416 changed the exception
message raised for base64 inputs whose length has a remainder of 1 when
divided by 4.  In these tests, we want to make sure that the underlying
exception message is passed through to allow debugging, but we don't
much care about the exact details, so change the input data to something
that raises an "Incorrect padding" exception on Python 3.8 as well as on
earlier versions.