Remove per-thread request debugging to filesystem
This was added in 2006, but it hasn't been something we've used at all
recently, and it's mostly superseded by Talisker logging.  Writing logs
into the code tree is problematic for charmed deployments, so I'd prefer
to avoid it.

If we need to bring this back for some reason, it should be behind a
configuration option that's only enabled for development.