Add initial appserver charm
There are still a lot more details of configuration to flesh out based
on lp:lp-production-configs, and some more things to set up such as
Nagios checks, an MTA, and cron jobs; but this gets us to the point
where it's possible to run a very minimal appserver and make local
requests to it, provided that you can separately arrange for a database
that it can connect to.

I've opted for the approach of generating `lazr.config` files based on
charm configuration.  This wasn't a completely obvious decision, since
it potentially means copying quite a lot of Launchpad's config schema
into charm `config.yaml` files.  However, as long as we're careful about
how we go about this (looking for patterns so that we can have a smaller
set of charm configuration keys), it has the benefit that it should be
much easier to avoid drift between our multiple environments.