Fix excessive logger cleanup in reset_logging
We reset the logging system between tests to avoid test isolation
issues.  However, we overdo it slightly by removing any loggers we don't
recognize from `Manager.loggerDict`.  This used to be OK, but as of
Python 3.7 the return value of `Logger.isEnabledFor` is cached, and the
mechanism for invalidating the cache when setting the log level relies
on being able to iterate over all loggers in `Manager.loggerDict`.

This only seemed to affect some webhook scheduling tests (depending on
test ordering, since it was a test isolation issue), but I suspect
that's mostly because they happen to be particularly pedantic about
their expected log output.