Reset logger propagation between tests
In a few places (`BuilddManager._setupLogger`, `TestGarbo.setUp`, and
`NumberCruncher._setupLogger`) we set up a logger in such a way that it
doesn't propagate unhandled messages to its parent logger.  We need to
reset such loggers between tests so that they don't cause isolation
failures with later tests.

I specifically excluded the "amqp", "brz", and "bzr" loggers from this
because `lp_sitecustomize` sets those to non-propagating, and we should
leave that alone.