Fetch files for private BPBs from librarian
This will allow us to lift the requirement for the source for private
binary package builds to be published before we can dispatch them.  We
were already using macaroon authentication for private source files due
to the `SnapBase` work last year; this just switches from having the
private PPA server do the authorization to having the librarian do it.

We now always fetch files for binary package builds using HTTPS, even
for public builds, which seems like a better idea in 2022.