Support parallel dispatch of a charm recipe for multiple series
`charmcraft` supports declaring that a single recipe should build for
multiple bases, which may include the case where bases differ only by
"channel" (i.e. distroseries), so don't reject build requests as
already-pending if an existing pending build of a recipe is for the same
architecture but a different series.

LP: #1986968