Use new OSError exception hierarchy in Python 3
This is generally less confusing than the catch-and-maybe-reraise idioms
we were previously using.

In one case the new code isn't precisely equivalent to the old code:
`lp.services.osutils.two_stage_kill` no longer suppresses `OSError`
exceptions other than the ones it explicitly catches (which should only
be possible here in the case of programming errors).  I think this was
probably a mistake in the previous code, since it's exactly the kind of
thing that's easy to do by mistake when using catch-and-maybe-reraise
idioms.