Replace `Cve.discoverer` with `discovered_by`
`Cve.discoverer` is a reference to a `Person`, but it doesn't really work with the real data in Ubuntu CVE Tracker.

In UCT, `Discovered-By` is a free-form field which usually contains name of one or more people or an organization name which is impossible to map to a `Person`.

We agreed that a plain text field would work better for this case.