Fix librarian URL generation for binary package builds
`LibraryFileAlias.https_url` is (confusingly) not the right thing to use
here, since it just takes the HTTP URL and unconditionally replaces its
scheme with "https".  We need to use `getURL()` instead, which uses
`https_url` for public files and constructs a URL on the appropriate
restricted domain for private files.