Fix race between branchChanged requests and branch scan jobs
If a `branchChanged` XML-RPC request happened to run at the same time as
a branch scan job for the same branch, then they may race to update the
same `Branch` row, causing one to have to roll back its transaction and
try again.  However, a bug in the `transaction` library meant that the
exception raised by the before-commit hook to gather job state was
incorrectly ignored, and as a result an `InFailedSqlTransaction`
exception was raised instead, which isn't retried.

Upgrade to `transaction==3.0.1` to fix this.

LP: #1883067