Fix the condition for dispatching tokens for private source files
If a librarian request has a token, then the librarian will only serve
restricted files.  In the case of source package release files, this
isn't quite synonymous with files in private archives: source files in
private archives may be unrestricted if the same source package is also
published in a public archive.  Only dispatch tokens for source files
that are actually restricted.