Allow setting launchpad_session.dbuser to None
In a charmed deployment, the PostgreSQL charm will provide us with a
full connection string including the user, and we won't want to override
that.  Allow setting `config.launchpad_session.dbuser` to None to
disable the arrangements that currently let us override the user on a
per-deployment basis.  This has no effect on existing deployments.