charm: Flesh out appserver configuration
This is enough to allow a charmed appserver deployment to have
equivalent configuration to each of our current production and staging
environments (apart from rabbitmq and memcached configuration, which
will be handled separately based on relations).

To avoid this being too overwhelming, I've stuck to only those bits of
configuration that are relevant to running the appserver.