Fix librarian-feed-swift's multiple-instance mode
The multiple-instance mode that I added in
https://code.launchpad.net/~cjwatson/launchpad/+git/launchpad/+merge/401127
turned out not to work quite right when we actually tried to use it on
production, because the instances raced with each other to list
directories and remove files from those directories in a way that could
result in exceptions while checking for recently-modified files.

Based on a change already applied to production by William Grant; I
added a corresponding test.