Include tokens in URLs returned by ArchiveAPI.translatePath
Files in the restricted librarian can only be accessed with some kind of
authentication token, either a `TimeLimitedToken` or a macaroon; without
adding something like that to the URLs returned by
`ArchiveAPI.translatePath`, there's no way to use it to access files in
private PPAs.  Macaroons don't offer any advantages in this case since
we can't constrain them in any interesting way, but a simple
`TimeLimitedToken` is fine.