Return proxied objects from various Bugs factory methods
`LaunchpadObjectFactory` issues `UnproxiedFactoryMethodWarning` when its
methods return objects not wrapped in a security proxy, since that tends
to result in tests that are less accurate simulations of production.

I made a number of tests in `TestBugSubscriptionFilter` set
`self.subscriber` (the logged-in user in this case) as the subscriber to
newly-created filters, since that allows them to access its attributes
without having to bypass security proxies.