Remove server feature scopes
We don't seem to use these in practice, and I can't really think of many
situations where it would be more convenient to do so rather than just
setting appropriate feature rules on each instance.  The only somewhat
relevant piece of configuration that we have right now is the `is_demo`
flag, but that isn't implemented using server scopes anyway and it
wouldn't be worth maintaining that code just for the sake of a single
boolean.