Use lazr.restful.ResourceJSONEncoder for mustache listings
I'd like to remove `simplejson` as a dependency of Launchpad; for the
most part it duplicates `json` from the standard library.  The only
exception I've run into is `simplejson.encoder.JSONEncoderForHTML`.
Fortunately, we already ran into this in `lazr.restful`, and so
`lazr.restful.ResourceJSONEncoder` is a sufficiently accurate emulation
of `simplejson.encoder.JSONEncoderForHTML` that we can use instead (it
does a few extra things as well, but that doesn't matter here).

Removing the bulk of trivial uses of `simplejson.dumps` and
`simplejson.loads` will be a separate branch.