Add ability to switch DB role after connecting
Our traditional strategy for deploying database credentials has been to
have a password for each database user we want to use, and to deploy
those to `.pgpass` files on each machine that needs to use each user.
This works, but it's quite manual and cumbersome.  With charmed
deployments it gets harder: each `db` relation can only communicate a
single set of credentials, which means that a script server would be
stuck having to have perhaps tens of relations representing the wide
variety of database users that the scripts it runs want to use: as well
as being ugly, I expect this would also make Juju very slow.

The natural approach with the PostgreSQL charm (or when emulating it
using `external-services`) is to set the `roles` attribute on the `db`
relation, which causes the relevant user to be made a member of that set
of roles.  The `SET ROLE` SQL command can then switch to that role.

I don't see a straightforward way to make these two strategies entirely
compatible, but it's simple enough to add a configuration flag that
selects which strategy we're using, so I've taken that approach.
`database.set_role_after_connecting` defaults to false which indicates
our traditional strategy; setting it to True causes us to use the `SET
ROLE` command after connecting instead, expecting the session user to
have been made a member of the target role.